requirements.txt
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-environ==0.11.2
requests==2.31.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.0
python-dateutil==2.8.2
psycopg==3.2.12
psycopg-binary==3.2.12
Pillow==10.1.0
gunicorn==21.2.0
django-storages==1.14.2
boto3==1.34.0
dj-database-url==2.1.0
whitenoise==6.6.0